+++ src/sdllib/shape.cc
@@ -1,6 +1,7 @@
 #include "sdllib/shape.h"
 
 #include <cstddef>
+#include <cstdint>
 #include <span>
 #include <utility>
 
@@ -18,7 +19,7 @@ int Extract_Shape_Count(const std::span<const std::byte> span) {
 
 void* Extract_Shape(const void* buffer, int shape) {
   auto* block = (ShapeBlock_Type*)buffer;
-  long offset;  // Offset of shape data, from start of block
+  uint32_t offset;  // Offset of shape data, from start of block
   char* bytebuf = (char*)buffer;
 
   /*
