+++ src/td/cdata.cc
@@ -1305,7 +1305,8 @@ const int16_t* TemplateTypeClass::Occupy_List(bool /*placement*/) const {
   int16_t* ptr;
   int index;
 
-  Mem_Copy(Get_Icon_Set_Map(Get_Image_Data()), map, static_cast<unsigned long>(Width) * Height);
+  Mem_Copy(Get_Icon_Set_Map(Get_Image_Data()), map,
+           static_cast<size_t>(Width) * Height);
 
   ptr = &_occupy[0];
   for (index = 0; index < Width * Height; index++) {
@@ -1394,7 +1395,8 @@ void TemplateTypeClass::Display(int x, int y, WindowNumberType window,
   x += WindowList[window][WINDOWX] << 3;
   y += WindowList[window][WINDOWY];
 
-  Mem_Copy(Get_Icon_Set_Map(Get_Image_Data()), map, static_cast<unsigned long>(Width) * Height);
+  Mem_Copy(Get_Icon_Set_Map(Get_Image_Data()), map,
+           static_cast<size_t>(Width) * Height);
 
   for (index = 0; index < w * h; index++) {
     if (map[index] != 0xFF) {
