+++ src/winvq/vqm32/palette.h
@@ -18,6 +18,8 @@
 
 #ifndef CNC_RED_ALERT_WINVQ_VQM32_PALETTE_H_
 #define CNC_RED_ALERT_WINVQ_VQM32_PALETTE_H_
+
+#include <cstdint>
 /****************************************************************************
  *
  *        C O N F I D E N T I A L -- W E S T W O O D  S T U D I O S
@@ -44,16 +46,16 @@
 extern "C" {
 #endif
 
-void __cdecl SetPalette(unsigned char* palette, long numbytes,
-                        unsigned long slowpal);
+void __cdecl SetPalette(unsigned char* palette, int32_t numbytes,
+                        uint32_t slowpal);
 void __cdecl ReadPalette(void* palette);
-void __cdecl SetDAC(long color, long red, long green, long blue);
-void __cdecl TranslatePalette(void* pal24, void* pal15, long numbytes);
+void __cdecl SetDAC(int32_t color, int32_t red, int32_t green, int32_t blue);
+void __cdecl TranslatePalette(void* pal24, void* pal15, int32_t numbytes);
 
 #ifdef __cplusplus
 }
 #endif
 
-void SortPalette(unsigned char* pal, long numcolors);
+void SortPalette(unsigned char* pal, int32_t numcolors);
 
 #endif  // CNC_RED_ALERT_WINVQ_VQM32_PALETTE_H_
